Add installation routine
The installation routine is supposed to make the installation process
more straightforward. The main tasks during installation are linking the
database and creating the first user. This is now possible in the
browser by calling install/
For security it is important to delete the installation directory
afterwards.

TODO: Therfore a usage of the software should not be possible while this
folder exists